==> R/01_stl-stack-fit_function.R <==
stl_stack_pred <- function(data, model_list, meta_model, horizon) {
  # Data preprocessing ------------------------------------------------------
  set.seed(1406)

  # stl decomposition
  stl_decomp <-
    stats::stl(
      x = ts(data = data$fires, frequency = 12),
      s.window = "per"
    )$time.series

  # create dataframes
  lag <- 5 # number of lags

  IMF <- list() # list to save dfs

  # create df for each imf with lags and inputs
  for (imf in seq(ncol(stl_decomp))) {
    IMF[[imf]] <-
      data.frame(TTR::lags(stl_decomp[, imf], lag))

    # rename columns
    names(IMF[[imf]]) <- c(
      "y",
      paste0("lag", seq(lag))
    )
  }

  # rename the list
  names(IMF) <- colnames(stl_decomp)

  # create training and test sets
  IMF_train <- list()
  IMF_test <- list()
  IMF_xtrain <- list()
  IMF_ytrain <- list()
  IMF_xtest <- list()
  IMF_ytest <- list()

  for (i in seq(IMF)) {
    n <- nrow(IMF[[i]])
    cut <- n - 12

    IMF_train[[i]] <- IMF[[i]][1:cut, ]
    IMF_test[[i]] <- tail(IMF[[i]], n - cut)

    IMF_xtrain[[i]] <- IMF_train[[i]][, -1]
    IMF_ytrain[[i]] <- IMF_train[[i]][, 1]

    IMF_xtest[[i]] <- IMF_test[[i]][, -1]
    IMF_ytest[[i]] <- IMF_test[[i]][, 1]
  }

  # create Observed dfs
  Obs <- TTR::lags(data$fires, lag)[, 1]
  Obs_train <- Obs[1:cut]
  Obs_test <- tail(Obs, n - cut)

  stl_df <- data.frame(data$fires, stl_decomp)
  colnames(stl_df) <- c("Obs", colnames(stl_decomp))

  # Training phase ----------------------------------------------------------
  # set random seed
  set.seed(1406)

  # set traincontrol
  control <- trainControl(
    method = "cv",
    number = 5
  )

  # list of training models
  model_list <- model_list

  # define objects
  IMF_model <- list()
  IMF_pred <- list()
  Params <- list()
  Importance <- list()
  k <- 1 # aux

  # training and predicting each IMF with each model
  for (imf in seq(IMF)) {
    Importance[[imf]] <- matrix(nrow = ncol(IMF_xtrain[[imf]]), ncol = length(model_list))
    colnames(Importance[[imf]]) <- model_list
    rownames(Importance[[imf]]) <- colnames(IMF_xtrain[[imf]])
    for (model in seq(model_list)) {
      # fitting
      IMF_model[[k]] <- train(
        y ~ .,
        data = IMF_train[[imf]],
        method = model_list[model],
        trControl = control,
        preProcess = c("center", "scale"),
        tuneLength = 5,
        trace = FALSE
      )

      # save hyperparameters
      Params[[k]] <- IMF_model[[k]]$bestTune

      # save variables importance
      Importance[[imf]][, model] <- varImp(IMF_model[[k]],
        scale = FALSE
      )$importance$Overall

      # prediction
      IMF_train_pred <- predict(IMF_model[[k]], IMF_train[[imf]])
      IMF_test_pred <- predict(IMF_model[[k]], IMF_test[[imf]])
      IMF_pred[[k]] <- data.frame(c(IMF_train_pred, IMF_test_pred))

      # print steps
      cat("\nModel: ", model_list[model], "\tc", imf, "\t",
        as.character(format.Date(Sys.time(), "%H:%M:%S")),
        sep = ""
      )

      # update k aux
      k <- k + 1
    }
  }

  names(Importance) <- names(IMF)

  # create a matrix combination
  combs <- matrix(nrow = length(model_list), ncol = length(IMF))
  colnames(combs) <- names(IMF)

  for (i in seq(model_list)) {
    for (j in seq(IMF)) {
      combs[i, j] <- i
    }
  }

  # Multi-step predictions ----

  # save IMF_model into a matrix
  model_matrix <- data.frame(matrix(IMF_model, ncol = length(IMF), byrow = FALSE))
  names(model_matrix) <- names(IMF)

  ## Recursive prediction

  # define objects
  step_stl_pred <- list()
  metrics_stl_train <- list()
  metrics_stl_test <- list()
  stack_model <- list()
  stack_params <- list()
  stack_pred <- list()
  metrics_stack_train <- list()
  metrics_stack_test <- list()
  predictions <- list()
  errors <- list()
  horizon <- horizon
  for (h in seq(length(horizon))) {{
    hrz <- horizon[h]
    PTRmo <- list()
    PTEmo <- list()
    Comp_train <- list()
    Comp_test <- list()
    k <- 1
  }

  cat("\nHorizon: ", hrz, "steps\n")

  for (m in 1:length(model_list)) {
    Comp_train[[m]] <- IMF_xtrain

    Comp_test[[m]] <- IMF_xtest

    PTRmo[[m]] <- matrix(ncol = length(IMF), nrow = nrow(IMF_xtrain[[1]]))
    PTEmo[[m]] <- matrix(ncol = length(IMF), nrow = nrow(IMF_xtest[[1]]))

    for (c in 1:length(Comp_train[[m]])) {
      if (hrz == 1) {
        # train
        PTRmo[[m]][, c] <- (predict(model_matrix[[m, c]], Comp_train[[m]][[c]]))

        # test
        PTEmo[[m]][, c] <- (predict(model_matrix[[m, c]], Comp_test[[m]][[c]]))
      } else {
        # train
        for (p in 1:cut) {
          if (p %% hrz != 1) {
            PTRmo[[m]][p, c] <- (predict(model_matrix[[m, c]], Comp_train[[m]][[c]][p, ]))
            if (hrz <= lag) {
              for (l in 1:(hrz - 1)) {
                Comp_train[[m]][[c]][p + l, l] <- PTRmo[[m]][p, c]
              }
            } else {
              for (l in 1:lag) {
                Comp_train[[m]][[c]][p + l, l] <- PTRmo[[m]][p, c]
              }
            }
          } else {
            Comp_train[[m]][[c]][p:cut, ] <- IMF_xtrain[[c]][p:cut, ]
            PTRmo[[m]][p, c] <- (predict(model_matrix[[m, c]], Comp_train[[m]][[c]][p, ]))
            if (hrz <= lag) {
              for (l in 1:(hrz - 1)) {
                Comp_train[[m]][[c]][p + l, l] <- PTRmo[[m]][p, c]
              }
            } else {
              for (l in 1:lag) {
                Comp_train[[m]][[c]][p + l, l] <- PTRmo[[m]][p, c]
              }
            }
          }
        }

        # test
        for (p in 1:(n - cut)) {
          if (p %% hrz != 1) {
            PTEmo[[m]][p, c] <- (predict(model_matrix[[m, c]], Comp_test[[m]][[c]][p, ]))
            if (hrz <= lag) {
              for (l in 1:(hrz - 1)) {
                Comp_test[[m]][[c]][p + l, l] <- PTEmo[[m]][p, c]
              }
            } else {
              for (l in 1:lag) {
                Comp_test[[m]][[c]][p + l, l] <- PTEmo[[m]][p, c]
              }
            }
          } else {
            Comp_test[[m]][[c]][p:(n - cut), ] <- IMF_xtest[[c]][p:(n - cut), ]
            PTEmo[[m]][p, c] <- (predict(model_matrix[[m, c]], Comp_test[[m]][[c]][p, ]))
            if (hrz <= lag) {
              for (l in 1:(hrz - 1)) {
                Comp_test[[m]][[c]][p + l, l] <- PTEmo[[m]][p, c]
              }
            } else {
              for (l in 1:lag) {
                Comp_test[[m]][[c]][p + l, l] <- PTEmo[[m]][p, c]
              }
            }
          }
        }
      }

      cat("Model: ", model_list[m], "\tc: ", c, "\t",
        (k / (length(model_list) * length(Comp_train[[m]]))) * 100, "%\n",
        sep = ""
      )

      k <- k + 1
    }
  }

  {
    metrics_stl_train[[h]] <- matrix(nrow = dim(combs)[1], ncol = 4)
    metrics_stl_test[[h]] <- matrix(nrow = dim(combs)[1], ncol = 4)
    colnames(metrics_stl_train[[h]]) <- c("i", "MAE", "MAPE", "RMSE")
    colnames(metrics_stl_test[[h]]) <- colnames(metrics_stl_train[[h]])
    rownames(metrics_stl_train[[h]]) <- model_list
    rownames(metrics_stl_test[[h]]) <- rownames(metrics_stl_train[[h]])

    step_stl_pred_train <- matrix(nrow = dim(IMF_xtrain[[1]])[1], ncol = dim(combs)[1])
    step_stl_pred_test <- matrix(nrow = dim(IMF_xtest[[1]])[1], ncol = dim(combs)[1])
    step_stl_pred[[h]] <- matrix(nrow = n, ncol = dim(combs)[1])
  }

  for (i in 1:dim(combs)[1]) {
    step_stl_pred_train[, i] <- rowSums(PTRmo[[i]])
    step_stl_pred_test[, i] <- rowSums(PTEmo[[i]])

    ### Avoiding negative values
    for (j in 1:dim(step_stl_pred_train)[1]) {
      if (step_stl_pred_train[j, i] < 0) {
        step_stl_pred_train[j, i] <- 0
      }
    }
    for (j in 1:dim(step_stl_pred_test)[1]) {
      if (step_stl_pred_test[j, i] < 0) {
        step_stl_pred_test[j, i] <- 0
      }
    }

    step_stl_pred[[h]][, i] <- c(
      step_stl_pred_train[, i],
      step_stl_pred_test[, i]
    )

    # metrics
    stl_step_mae_train <- MAE(step_stl_pred_train[, i], Obs_train)
    stl_step_mape_train <- mape(step_stl_pred_train[, i], Obs_train)
    stl_step_rmse_train <- RMSE(step_stl_pred_train[, i], Obs_train)

    stl_step_mae_test <- MAE(step_stl_pred_test[, i], Obs_test)
    stl_step_mape_test <- mape(step_stl_pred_test[, i], Obs_test)
    stl_step_rmse_test <- RMSE(step_stl_pred_test[, i], Obs_test)

    metrics_stl_train[[h]][i, ] <- c(
      i,
      stl_step_mae_train,
      stl_step_mape_train,
      stl_step_rmse_train
    )
    metrics_stl_test[[h]][i, ] <- c(
      i,
      stl_step_mae_test,
      stl_step_mape_test,
      stl_step_rmse_test
    )
  }
  step_stl_pred[[h]] <- cbind(Obs, step_stl_pred[[h]])
  colnames(step_stl_pred[[h]]) <- c("Obs", model_list)

  # Stacking ----------------------------------------------------------------
  cat("\nStacking\n")
  stack_data <- step_stl_pred[[h]]
  colnames(stack_data) <- c("y", model_list)

  stack_data_train <- stack_data[1:cut, ]
  stack_data_test <- tail(stack_data, n - cut)

  meta_model <- meta_model

  stack_pred_train <- matrix(nrow = nrow(stack_data_train), ncol = 1)
  stack_pred_test <- matrix(nrow = nrow(stack_data_test), ncol = 1)
  stack_pred[[h]] <- matrix(nrow = n, ncol = 1)

  metrics_stack_train[[h]] <- matrix(nrow = 1, ncol = 3)
  metrics_stack_test[[h]] <- matrix(nrow = 1, ncol = 3)
  colnames(metrics_stack_train[[h]]) <- c("MAE", "MAPE", "RMSE")
  colnames(metrics_stack_test[[h]]) <- colnames(metrics_stack_train[[h]])

  stack_model[[h]] <- train(
    y ~ .,
    data = stack_data_train,
    method = meta_model,
    trControl = control,
    preProcess = c("center", "scale"),
    trace = FALSE
  )

  # save hyperparameters
  stack_params[[h]] <- stack_model[[h]]$bestTune

  # prediction
  stack_train_pred <- predict(stack_model[[h]], stack_data_train)
  stack_test_pred <- predict(stack_model[[h]], stack_data_test)
  stack_pred[[h]] <- c(stack_train_pred, stack_test_pred)

  # metrics
  stack_mae_train <- MAE(stack_train_pred, Obs_train)
  stack_mape_train <- mape(stack_train_pred, Obs_train)
  stack_rmse_train <- RMSE(stack_train_pred, Obs_train)

  stack_mae_test <- MAE(stack_test_pred, Obs_test)
  stack_mape_test <- mape(stack_test_pred, Obs_test)
  stack_rmse_test <- RMSE(stack_test_pred, Obs_test)

  metrics_stack_train[[h]] <- c(
    stack_mae_train,
    stack_mape_train,
    stack_rmse_train
  )
  metrics_stack_test[[h]] <- c(
    stack_mae_test,
    stack_mape_test,
    stack_rmse_test
  )

  predictions[[h]] <- cbind(Obs, stack_pred[[h]], step_stl_pred[[h]][, -1])

  errors[[h]] <- matrix(ncol = ncol(predictions[[h]]) - 1, nrow = n)
  colnames(errors[[h]]) <- colnames(predictions[[h]][, -1])

  for (error in seq(ncol(errors[[h]]))) {
    errors[[h]][, error] <- (predictions[[h]][, 1] - predictions[[h]][, 1 + error])
  }}
  names(step_stl_pred) <- paste0(horizon, "-step")
  names(metrics_stl_train) <- names(step_stl_pred)
  names(metrics_stl_test) <- names(step_stl_pred)
  names(metrics_stack_train) <- names(step_stl_pred)
  names(metrics_stack_test) <- names(step_stl_pred)
  names(stack_pred) <- names(step_stl_pred)
  names(predictions) <- names(step_stl_pred)
  names(errors) <- names(step_stl_pred)

  results <- list(
    Predictions = predictions,
    Decomp_Metrics = metrics_stl_test,
    STACK_Metrics = metrics_stack_test,
    Hyperparameters = Params,
    Stack_params = stack_params,
    Var_Importance = Importance,
    Errors = errors,
    STL = stl_df
  )

  return(results)
}
